Fix filtering bug
