+++ src/components/elements/tableFilters/filters/FilterItem.tsx
@@ -8,9 +8,8 @@ import PickListWrapper from './items/PickListWrapper';
 import TableFilterItemSelect from './items/TableFilterItemSelect';
 
 import { TableFilterItemSelectorProps } from './items/types';
-import { FilterType, SelectElementVariant } from '@/modules/dataFetching/types';
+import { FilterType } from '@/modules/dataFetching/types';
 import { localResolvePickList } from '@/modules/form/util/formUtil';
-import { PickListOption } from '@/types/gqlTypes';
 
 const TableFilterItemSelector = ({
   variant = 'select',
@@ -20,19 +19,6 @@ const TableFilterItemSelector = ({
   if (variant === 'select') return <TableFilterItemSelect {...props} />;
 
   return <TableFilterItemSelect {...props} />;
-};
-
-const selectDefaultFilterVariant = (filter: {
-  variant?: SelectElementVariant;
-  pickListOptions?: PickListOption[];
-  multi?: boolean;
-}): SelectElementVariant => {
-  return (
-    filter.variant ||
-    (filter.pickListOptions && filter.pickListOptions.length < 3 && filter.multi
-      ? 'checkboxes'
-      : 'select')
-  );
 };
 
 export interface TableFilterItemProps<T> {
@@ -87,10 +73,9 @@ const TableFilterItem = <T,>({
             filter.pickListOptions ||
             localResolvePickList(filter.enumType, true) ||
             [];
-          const variant = selectDefaultFilterVariant(filter);
           return (
             <TableFilterItemSelector
-              variant={variant}
+              variant={filter.variant}
               options={options}
               value={filter.multi ? value || [] : value}
               onChange={onChange}
@@ -100,10 +85,9 @@ const TableFilterItem = <T,>({
         }
 
         if (filter.type === 'local_picklist') {
-          const variant = selectDefaultFilterVariant(filter);
           return (
             <TableFilterItemSelector
-              variant={variant}
+              variant={filter.variant}
               options={filter.pickListOptions}
               value={filter.multi ? value || [] : value}
               onChange={onChange}
